refactor: 공지사항 컴포넌트 리팩토링 및 import 업데이트 #83
- 일관성을 위해 공지사항 컴포넌트 파일 이름 변경 및 관련 import 업데이트
- AnnouncementSearchDialog와 AnnouncementSearchSection의 레이아웃과 스타일 개선
- Tailwind safelist에 'border-t-gray-900' 추가
