+++ tailwind.config.js
@@ -10,6 +10,7 @@ module.exports = {
     'mx-auto',
     'min-w-0',
     'flex-1',
+    'border-t-gray-900',
     'border-transparent',
     'border-b',
     'border-r',
